Fix error in qc.df() when no query is active
Previously, the examples/app-database.R would shown an error on
startup because the initial query was "", which was then sent
as a SQL query to RSQLite. The get_lazy_data code path accounted
for the "" query, so we decided to make the eager code path just
call the lazy code path, then collect().

Also fixed a formatting issue with the table.